pyproject.toml
[project]
name = "marknotion"
version = "0.1.0"
description = "Bidirectional Markdown ↔ Notion blocks converter"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [{ name = "Libo" }]
keywords = ["markdown", "notion", "converter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marknotion"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
